add GameActivity and layout
