+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Fri Sep 17 09:52:27 CEST 2021
+gradle.version=7.0.2
