+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.BlackJackAppX">
         <activity
+            android:name=".GameActivity"
+            android:exported="true" />
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
