+++ app/src/main/java/com/example/blackjackappx/MainActivity.kt
@@ -8,20 +8,8 @@ import com.example.blackjackappx.entities.User
 
 class MainActivity : AppCompatActivity() {
 
-    lateinit var btnLogout: Button
-    lateinit var tvPlayerStack: TextView
-    lateinit var tvPlayerName: TextView
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
-        btnLogout = findViewById(R.id.btn_logout)
-        tvPlayerStack = findViewById(R.id.tv_stack_size)
-        tvPlayerName = findViewById(R.id.tvPlayerName)
-        val userOne: User = User("Lisa", "123")
-        tvPlayerName.setText(userOne.userName)
-        tvPlayerStack.setText(userOne.userStack.toString())
-
     }
 }
