+++ app/src/main/res/layout/activity_main.xml
@@ -7,112 +7,4 @@
     android:background="@drawable/mat_3"
     tools:context=".MainActivity">
 
-    <Button
-        android:id="@+id/btn_logout"
-        android:layout_width="110dp"
-        android:layout_height="70dp"
-        android:layout_marginEnd="5dp"
-        android:layout_marginBottom="5dp"
-        android:backgroundTint="#1A1C1C"
-        android:text="@string/logoutText"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <Button
-        android:id="@+id/btn_draw"
-        android:layout_width="0dp"
-        android:layout_height="90dp"
-        android:layout_marginStart="5dp"
-        android:layout_marginEnd="5dp"
-        android:layout_marginBottom="192dp"
-        android:backgroundTint="#4AEF51"
-        android:text="@string/drawText"
-        android:textColor="#263238"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/btn_hold"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <Button
-        android:id="@+id/btn_hold"
-        android:layout_width="0dp"
-        android:layout_height="90dp"
-        android:layout_marginBottom="192dp"
-        android:backgroundTint="#F44A4A"
-        android:text="@string/holdText"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/btn_bet"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/btn_draw" />
-
-    <Button
-        android:id="@+id/btn_bet"
-        android:layout_width="120dp"
-        android:layout_height="90dp"
-        android:layout_marginStart="5dp"
-        android:layout_marginEnd="5dp"
-        android:layout_marginBottom="192dp"
-        android:backgroundTint="#3DA2CF"
-        android:text="@string/betText"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/btn_hold" />
-
-    <TextView
-        android:id="@+id/tv_stack_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginBottom="10dp"
-        android:text="@string/userStackText"
-        android:textSize="24sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <TextView
-        android:id="@+id/tv_stack_size"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="4dp"
-        android:layout_marginBottom="10dp"
-        android:text="@string/playerStack"
-        android:textSize="24sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/tv_stack_text" />
-
-    <EditText
-        android:id="@+id/et_bet"
-        android:layout_width="102dp"
-        android:layout_height="57dp"
-        android:layout_marginTop="4dp"
-        android:layout_marginEnd="16dp"
-        android:ems="10"
-        android:inputType="textPersonName"
-        android:text="@string/betSize2"
-        android:textSize="34sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btn_bet" />
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginBottom="10dp"
-        android:text="@string/playerText1"
-        android:textSize="24sp"
-        app:layout_constraintBottom_toTopOf="@+id/btn_draw"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <TextView
-        android:id="@+id/tvPlayerName"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="2dp"
-        android:layout_marginBottom="10dp"
-        android:text="@string/playerName"
-        android:textSize="24sp"
-        app:layout_constraintBottom_toTopOf="@+id/btn_hold"
-        app:layout_constraintStart_toEndOf="@+id/textView" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
