+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Thu Sep 16 13:15:24 CEST 2021
-sdk.dir=C\:\\Users\\jonas\\AppData\\Local\\Android\\Sdk
+#Fri Sep 17 09:52:19 CEST 2021
+sdk.dir=C\:\\Users\\ziggi\\AppData\\Local\\Android\\Sdk
